Created effect, reducer, selector for users
I created an effect that retrives the users from BE and compares their credentials with the ones introduced by the user.

I created a reducer that when it's activated sets "isUserValid" on true or false based on the effect above.

I created a selector that retrives the result of "isUserValid"